requirements.txt
djangorestframework-simplejwt
django-cors-headers
